feat(uorm): 引入 SQLite 支持并优化模板引擎
- 新增 SQLite 数据库驱动支持，包括连接、查询、执行等完整功能
- 实现 SQLite 特有的连接池和事务管理机制
- 添加 rusqlite 依赖并配置可选特性支持
- 更新模板引擎，重命名 for 标签为 foreach 并调整属性名称
- 移除模板缓存的主动清除接口，简化缓存管理逻辑
- 新增多个集成测试用例，覆盖 SQLite 的基本 CRUD 和事务操作
- 调整默认特性，优先启用 SQLite 而非 MySQL

+++ src/udbc/mysql/value_codec.rs
@@ -16,15 +16,14 @@ pub fn from_mysql_value(v: &MyValue) -> Value {
             } else {
                 let dt = NaiveDate::from_ymd_opt(*y as i32, *m as u32, *d as u32)
                     .unwrap()
-                    .and_hms_micro_opt(*h as u32, *min as u32, *s as u32, *micro as u32)
+                    .and_hms_micro_opt(*h as u32, *min as u32, *s as u32, *micro)
                     .unwrap();
                 Value::DateTime(dt)
             }
         }
         MyValue::Time(is_neg, days, h, min, s, micro) => {
             let total_h = *days * 24 + (*h as u32);
-            let t = NaiveTime::from_hms_micro_opt(total_h, *min as u32, *s as u32, *micro as u32)
-                .unwrap();
+            let t = NaiveTime::from_hms_micro_opt(total_h, *min as u32, *s as u32, *micro).unwrap();
             if *is_neg {
                 Value::Str(format!("-{}", t))
             } else {
